--- benchmark/api/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@verdaccio/benckmark-api",
-  "version": "1.0.0",
-  "private": "true",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "ts-node index.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@verdaccio/node-api": "workspace:6.0.0-6-next.16",
-    "autocannon": "7.3.0",
-    "ts-node": "10.0.0"
-  }
-}
