--- benchmark/api/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.reference.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./build"
-  },
-  "files": ["index.ts"]
-}
