--- benchmark/hyperfine/fixtures/empty.pkg.json
@@ -1,12 +1,0 @@
-{
-  "name": "empty-project",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
