--- benchmark/hyperfine/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@verdaccio/benckmark-hyperfine",
-  "version": "1.0.0",
-  "private": "true",
-  "description": "",
-  "main": "src/init.ts",
-  "scripts": {
-    "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "clipanion": "3.0.0-rc.12"
-  },
-  "devDependencies": {
-    "@verdaccio/node-api": "workspace:6.0.0-6-next.16",
-    "node-fetch": "2.6.1"
-  }
-}
