+++ package.json
@@ -72,6 +72,7 @@
     "@typescript-eslint/parser": "4.13.0",
     "@verdaccio/types": "workspace:*",
     "@verdaccio/ui-theme": "workspace:*",
+    "@verdaccio/benchmark": "workspace:*",
     "@verdaccio/eslint-config": "workspace:*",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
@@ -135,8 +136,8 @@
     "test:e2e:cli": "pnpm test --filter ...@verdaccio/e2e-cli",
     "test:e2e:ui": "pnpm test --filter ...@verdaccio/e2e-ui",
     "start": "concurrently --kill-others \"pnpm _start:server\" \"pnpm _start:web\"",
-    "benchmark:hyper": "ts-node benchmark/hyperfine/index.ts --report ./hyper-results.json",
-    "benchmark:api": "ts-node benchmark/api/index.ts",
+    "benchmark:hyper": "verdaccio-benchmark hyper -r ./hyper-results.json",
+    "benchmark:api": "verdaccio-benchmark api",
     "start:watch": "concurrently --kill-others \"pnpm _build:watch\" \"pnpm _start:server\" \"pnpm _debug:reload\"",
     "_build:watch": "pnpm run --parallel watch --filter ./packages",
     "_start:server": "node packages/verdaccio/debug/bootstrap.js --listen 8000",
