+++ packages/tools/benchmark/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../.babelrc"
+}
